* for DX11 I now use shader reflection to get the input layout (http://takinginitiative.net/2011/12/11/directx-1011-basic-shader-reflection-automatic-input-layout-creation/) * changed how the init params are passed on to resources by the resource manager * fixed build

+++ src/Core/FileSystem/File_windows.cpp
@@ -49,7 +49,7 @@ namespace Oak3D
 			struct __stat64 fileStat; 
 			if(_stat64( filepath.c_str(), &fileStat ))
 				return 0; 
-			return fileStat.st_size;
+			return (unsigned long)fileStat.st_size;
 		}
 
 		unsigned long File::Size()
